SenSa/requirements.txt
asgiref==3.11.1
black==25.11.0
cfgv==3.4.0
click==8.1.8
distlib==0.4.0
django==4.2.29
django-cors-headers>=4.3
djangorestframework==3.16.1
djangorestframework-simplejwt>=5.3
filelock==3.24.2
identify==2.6.15
mypy-extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.4.0
pre-commit==4.3.0
python-discovery==1.2.0
python-dotenv>=1.0
pytokens==0.4.1
pyyaml==6.0.3
ruff==0.15.6
setuptools==58.0.4
sqlparse==0.5.5
tomli==2.4.0
typing-extensions==4.15.0
virtualenv==21.2.0